requirements.txt
flask
gunicorn
torch>=1.7.0
torchvision
Pillow
numpy
opencv-python
pyyaml
tqdm
requests
pandas
scipy
